Many modifications that sadly I didn't track properly
I have been working on this on and off. The main use was to provide
an easy way for my in-laws to copy their photos from their camera to
their computer.

The final goal was to have a graphical UI display what's being done,
but I never got around properly finish it.

Anyway I had this version at home, which seems more advanced that what
I put on github, so there it is. The code is almost a year old.